server/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# WebSocket support
websockets==12.0

# File handling and utilities
python-multipart==0.0.6
aiofiles==23.2.1

# Environment and configuration
python-dotenv==1.0.0

# HTTP client for health checks
aiohttp==3.9.1

# Hardware monitoring
pythonnet==3.0.3
HardwareMonitor==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort

# Security
cryptography==41.0.7

# For structured logging
structlog

# For system metrics in health check
psutil
